requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
prefect>=3.1
requests==2.32.0
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
sqlalchemy==2.0.23
mlflow==2.8.1
pydantic==2.5.2
pytest==7.4.3
loguru==0.7.2
streamlit==1.28.1 